Add files via upload
In this Problem, you have to write an application in Python 3.7 that maps IPL franchises and players as per the below guidelines. 
Assume that you are a news reporter and you want to map which players have been associated with a franchise (either in the past or present). For this you need to have some system of storing these players and the franchise they have been with. Assume that you have a list of N franchises and M players. For the sake of this assignment, let us assume that a particular player could be associated with only two franchises at max. 

Model the following problem as a graph-based problem. Clearly state how the vertices and edges can be modelled such that this graph can be used to answer the following queries efficiently. 
1. List the unique franchises and players the reporter has collected in the system. 
2. For a particular player, help the reporter recollect the franchises he has represented. 
3. For a particular franchise, list the players that have been associated with it (past or present). 
4. Identify if two players are franchise buddies. Player A and Player B are considered to be franchise buddies if they have been associated with the same franchise (not necessarily at the same time or in the same year) 
5. Can two players A and B be connected such that there exists another player C where A and C are franchise buddies and C and B are franchise buddies. 
6. Perform an analysis for the questions above and give the running time in terms of n.